test fails: AggregateFunction from wrong place
