+++ src/common/src/queries/event_segmentation/mod.rs
@@ -1,6 +1,8 @@
 use chrono::DateTime;
 use chrono::Utc;
-use datafusion_expr::AggregateFunction;
+
+// use datafusion_expr::AggregateFunction;
+use crate::queries::AggregateFunction;
 
 pub mod logical_plan_builder;
