+++ src/platform/src/queries/mod.rs
@@ -119,94 +119,94 @@ impl TryInto<common::queries::TimeIntervalUnit> for TimeIntervalUnit {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum AggregateFunction {
-    /// count
-    Count,
-    /// sum
-    Sum,
-    /// min
-    Min,
-    /// max
-    Max,
-    /// avg
-    Avg,
-    /// Approximate aggregate function
-    ApproxDistinct,
-    /// array_agg
-    ArrayAgg,
-    /// Variance (Sample)
-    Variance,
-    /// Variance (Population)
-    VariancePop,
-    /// Standard Deviation (Sample)
-    Stddev,
-    /// Standard Deviation (Population)
-    StddevPop,
-    /// Covariance (Sample)
-    Covariance,
-    /// Covariance (Population)
-    CovariancePop,
-    /// Correlation
-    Correlation,
-    /// Approximate continuous percentile function
-    ApproxPercentileCont,
-    /// ApproxMedian
-    ApproxMedian,
-}
+// #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Hash, Serialize, Deserialize)]
+// #[serde(rename_all = "camelCase")]
+// pub enum AggregateFunction {
+//     /// count
+//     Count,
+//     /// sum
+//     Sum,
+//     /// min
+//     Min,
+//     /// max
+//     Max,
+//     /// avg
+//     Avg,
+//     /// Approximate aggregate function
+//     ApproxDistinct,
+//     /// array_agg
+//     ArrayAgg,
+//     /// Variance (Sample)
+//     Variance,
+//     /// Variance (Population)
+//     VariancePop,
+//     /// Standard Deviation (Sample)
+//     Stddev,
+//     /// Standard Deviation (Population)
+//     StddevPop,
+//     /// Covariance (Sample)
+//     Covariance,
+//     /// Covariance (Population)
+//     CovariancePop,
+//     /// Correlation
+//     Correlation,
+//     /// Approximate continuous percentile function
+//     ApproxPercentileCont,
+//     /// ApproxMedian
+//     ApproxMedian,
+// }
 
-impl TryInto<datafusion::physical_plan::aggregates::AggregateFunction> for &AggregateFunction {
-    type Error = PlatformError;
+// impl TryInto<datafusion::physical_plan::aggregates::AggregateFunction> for &AggregateFunction {
+//     type Error = PlatformError;
 
-    fn try_into(
-        self,
-    ) -> std::result::Result<datafusion::physical_plan::aggregates::AggregateFunction, Self::Error>
-    {
-        Ok(match self {
-            AggregateFunction::Count => {
-                datafusion::physical_plan::aggregates::AggregateFunction::Count
-            }
-            AggregateFunction::Sum => datafusion::physical_plan::aggregates::AggregateFunction::Sum,
-            AggregateFunction::Min => datafusion::physical_plan::aggregates::AggregateFunction::Min,
-            AggregateFunction::Max => datafusion::physical_plan::aggregates::AggregateFunction::Max,
-            AggregateFunction::Avg => datafusion::physical_plan::aggregates::AggregateFunction::Avg,
-            AggregateFunction::ApproxDistinct => {
-                datafusion::physical_plan::aggregates::AggregateFunction::ApproxDistinct
-            }
-            AggregateFunction::ArrayAgg => {
-                datafusion::physical_plan::aggregates::AggregateFunction::ArrayAgg
-            }
-            AggregateFunction::Variance => {
-                datafusion::physical_plan::aggregates::AggregateFunction::Variance
-            }
-            AggregateFunction::VariancePop => {
-                datafusion::physical_plan::aggregates::AggregateFunction::VariancePop
-            }
-            AggregateFunction::Stddev => {
-                datafusion::physical_plan::aggregates::AggregateFunction::Stddev
-            }
-            AggregateFunction::StddevPop => {
-                datafusion::physical_plan::aggregates::AggregateFunction::StddevPop
-            }
-            AggregateFunction::Covariance => {
-                datafusion::physical_plan::aggregates::AggregateFunction::Covariance
-            }
-            AggregateFunction::CovariancePop => {
-                datafusion::physical_plan::aggregates::AggregateFunction::CovariancePop
-            }
-            AggregateFunction::Correlation => {
-                datafusion::physical_plan::aggregates::AggregateFunction::Correlation
-            }
-            AggregateFunction::ApproxPercentileCont => {
-                datafusion::physical_plan::aggregates::AggregateFunction::ApproxPercentileCont
-            }
-            AggregateFunction::ApproxMedian => {
-                datafusion::physical_plan::aggregates::AggregateFunction::ApproxMedian
-            }
-        })
-    }
-}
+//     fn try_into(
+//         self,
+//     ) -> std::result::Result<datafusion::physical_plan::aggregates::AggregateFunction, Self::Error>
+//     {
+//         Ok(match self {
+//             AggregateFunction::Count => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::Count
+//             }
+//             AggregateFunction::Sum => datafusion::physical_plan::aggregates::AggregateFunction::Sum,
+//             AggregateFunction::Min => datafusion::physical_plan::aggregates::AggregateFunction::Min,
+//             AggregateFunction::Max => datafusion::physical_plan::aggregates::AggregateFunction::Max,
+//             AggregateFunction::Avg => datafusion::physical_plan::aggregates::AggregateFunction::Avg,
+//             AggregateFunction::ApproxDistinct => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::ApproxDistinct
+//             }
+//             AggregateFunction::ArrayAgg => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::ArrayAgg
+//             }
+//             AggregateFunction::Variance => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::Variance
+//             }
+//             AggregateFunction::VariancePop => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::VariancePop
+//             }
+//             AggregateFunction::Stddev => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::Stddev
+//             }
+//             AggregateFunction::StddevPop => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::StddevPop
+//             }
+//             AggregateFunction::Covariance => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::Covariance
+//             }
+//             AggregateFunction::CovariancePop => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::CovariancePop
+//             }
+//             AggregateFunction::Correlation => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::Correlation
+//             }
+//             AggregateFunction::ApproxPercentileCont => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::ApproxPercentileCont
+//             }
+//             AggregateFunction::ApproxMedian => {
+//                 datafusion::physical_plan::aggregates::AggregateFunction::ApproxMedian
+//             }
+//         })
+//     }
+// }
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Hash, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
