+++ src/platform/src/reports/provider_impl.rs
@@ -8,6 +8,7 @@ use super::CreateReportRequest;
 use super::Provider;
 use super::Report;
 use crate::Context;
+use crate::PlatformError;
 
 pub struct ProviderImpl {
     prov: Arc<dyn ReportsProvider>,
@@ -19,37 +20,37 @@ impl ProviderImpl {
     }
 }
 
-#[async_trait]
-impl Provider for ProviderImpl {
-    async fn create(
-        &self,
-        ctx: Context,
-        organization_id: u64,
-        project_id: u64,
-        request: CreateReportRequest,
-    ) -> Result<Report> {
-        ctx.check_project_permission(
-            organization_id,
-            project_id,
-            ProjectPermission::ManageReports,
-        )?;
+// #[async_trait]
+// impl Provider for ProviderImpl {
+//     async fn create(
+//         &self,
+//         ctx: Context,
+//         organization_id: u64,
+//         project_id: u64,
+//         request: CreateReportRequest,
+//     ) -> Result<Report, PlatformError> {
+//         ctx.check_project_permission(
+//             organization_id,
+//             project_id,
+//             ProjectPermission::ManageReports,
+//         )?;
 
-        let report = self
-            .prov
-            .create(
-                organization_id,
-                project_id,
-                metadata::reports::CreateReportRequest {
-                    created_by: ctx.account_id.unwrap(),
-                    tags: request.tags,
-                    name: request.name,
-                    description: request.description,
-                    typ: request.typ,
-                    query: request.query,
-                },
-            )
-            .await?;
+//         let report = self
+//             .prov
+//             .create(
+//                 organization_id,
+//                 project_id,
+//                 metadata::reports::CreateReportRequest {
+//                     created_by: ctx.account_id.unwrap(),
+//                     tags: request.tags,
+//                     name: request.name,
+//                     description: request.description,
+//                     typ: request.typ,
+//                     query: request.query,
+//                 },
+//             )
+//             .await?;
 
-        report.try_into()
-    }
-}
+//         report.try_into()
+//     }
+// }
